+++ Makefile
@@ -1,13 +1,6 @@
 MODULE_big = fio
 
 OBJS = src/fio.o       \
-       src/renamefile.o \
-       src/removefile.o \
-       src/writefile.o \
-       src/readfile.o  \
-       src/readdir.o   \
-       src/mkdir.o     \
-       src/chmod.o     \
        src/utils.o
 
 EXTENSION = fio
