+++ src/fio.h
@@ -54,14 +54,13 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "utils.h"
 
-Datum fio_writefile(PG_FUNCTION_ARGS);
-Datum fio_readfile(PG_FUNCTION_ARGS);
-Datum fio_readdir(PG_FUNCTION_ARGS);
-Datum fio_mkdir(PG_FUNCTION_ARGS);
-Datum fio_chmod(PG_FUNCTION_ARGS);
-Datum fio_removefile(PG_FUNCTION_ARGS);
-Datum fio_renamefile(PG_FUNCTION_ARGS);
 #endif
+
+#define BUFFER_SIZE 1024
+
+struct dircontext {
+    DIR *dir;
+};
 
 #ifndef FALSE
 #define FALSE   (0)
